refactor: Apis 폴더 네이밍 및 위치 수정
